Fix issue101 - Localize Thank you page
